+++ src/functions/mod.rs
@@ -1,5 +1,5 @@
 pub mod list;
 pub mod position;
+pub mod set_property;
 pub mod show;
 pub mod show_output;
-pub mod set_property;