+++ src/wlr_client/output/mod.rs
@@ -1,6 +1,6 @@
 pub(crate) mod config_writer;
 pub(crate) mod configs;
+pub(crate) mod heads;
 pub(crate) mod output_manager;
 pub(crate) mod wlr_head;
 pub(crate) mod wlr_mode;
-pub(crate) mod heads;