+++ tests/integration/tests/mod.rs
@@ -4,7 +4,10 @@ mod position;
 mod set;
 mod show;
 
-fn find_output<'a>(outputs: &'a [crate::swaymsg::Output], name: &str) -> &'a crate::swaymsg::Output {
+fn find_output<'a>(
+    outputs: &'a [crate::swaymsg::Output],
+    name: &str,
+) -> &'a crate::swaymsg::Output {
     outputs
         .iter()
         .find(|o| o.name == name)